backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
anyio==3.7.1  # Required by FastAPI

# Supabase - Fully compatible version set for Python 3.13
supabase==2.9.0
postgrest==0.17.1
httpx==0.25.2
realtime==2.0.5
storage3==0.8.0
gotrue==2.9.1
supafunc==0.5.1

# Data processing
pandas>=2.0.0
numpy>=1.24.0
python-dateutil==2.8.2

# Data sources
yfinance>=0.2.48
requests==2.31.0

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
statsmodels==0.14.0

# Feature engineering
ta==0.11.0  # Technical Analysis library (RSI, MACD, Bollinger, etc.)

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytz==2023.3.post1
schedule==1.2.0

# Monitoring and logging
loguru==0.7.2

# Testing
pytest==7.4.3
